backend/requirements.txt
Flask==2.2.5
sentence-transformers==2.2.2
langchain==0.0.180
chromadb==0.3.21
requests==2.28.1
python-docx==0.8.11
PyMuPDF==1.20.0
transformers==4.46.3
torch==2.4.1
